a few helpful config patches

+++ src/main/java/com/github/standobyte/jojo/config/core/types/ConfigFloat.java
@@ -9,7 +9,7 @@ import net.minecraft.network.codec.ByteBufCodecs;
 public class ConfigFloat extends ConfigOption<Float> {
 	protected DefaultedValue.Float value;
 
-	public ConfigFloat(int defaultValue) {
+	public ConfigFloat(float defaultValue) {
 		super(PrimitiveCodec.FLOAT, ByteBufCodecs.FLOAT);
 		this.value = new DefaultedValue.Float(defaultValue);
 	}
